moved login/logout callbacks to profile module; renamed oauth utils to auth
