+++ geonode/people/models.py
@@ -18,21 +18,31 @@
 #
 #########################################################################
 
-from uuid import uuid4
+from uuid import uuid1, uuid4
+import logging
+import traceback
 
 from django.db import models
 from django.utils.translation import ugettext_lazy as _
 from django.core.urlresolvers import reverse
 from django.contrib.auth.models import AbstractUser, UserManager
+from django.contrib.auth.signals import user_logged_in, user_logged_out
 from django.contrib.sites.models import Site
 from django.db.models import signals
 from django.conf import settings
 
 from taggit.managers import TaggableManager
 
+from geonode.decorators import on_ogc_backend
 from geonode.base.enumerations import COUNTRIES
+from geonode.base.auth import (get_or_create_token,
+                               delete_old_tokens,
+                               set_session_token,
+                               remove_session_token)
 from geonode.groups.models import GroupProfile
 # from geonode.notifications_helper import send_notification
+
+from geonode import geoserver
 
 from allauth.account.signals import user_signed_up
 from allauth.socialaccount.signals import social_account_added
@@ -43,6 +53,8 @@ from .signals import (update_user_email_addresses,
                       notify_admins_new_signup)
 from .languages import LANGUAGES
 from .timezones import TIMEZONES
+
+logger = logging.getLogger(__name__)
 
 
 class ProfileUserManager(UserManager):
@@ -237,3 +249,38 @@ user_signed_up.connect(
     weak=False
 )
 signals.post_save.connect(profile_post_save, sender=Profile)
+
+
+@on_ogc_backend(geoserver.BACKEND_PACKAGE)
+def do_login(sender, user, request, **kwargs):
+    """
+    Take action on user login. Generate a new user access_token to be shared
+    with GeoServer, and store it into the request.session
+    """
+    if user and user.is_authenticated():
+        token = None
+        try:
+            token = get_or_create_token(user)
+        except BaseException:
+            u = uuid1()
+            token = u.hex
+            tb = traceback.format_exc()
+            logger.debug(tb)
+
+        set_session_token(request.session, token)
+
+
+@on_ogc_backend(geoserver.BACKEND_PACKAGE)
+def do_logout(sender, user, request, **kwargs):
+    if 'access_token' in request.session:
+        try:
+            delete_old_tokens(user)
+        except BaseException:
+            tb = traceback.format_exc()
+            logger.debug(tb)
+        remove_session_token(request.session)
+        request.session.modified = True
+
+
+user_logged_in.connect(do_login)
+user_logged_out.connect(do_logout)
